# openapi.yaml
openapi: 3.0.1
info:
  title: Laine Déco API
  version: 1.0.0
  description: |
    Documentation complète de l'API Laine Déco - Plateforme de vente de laine et accessoires

    ## 🔐 Authentification
    Utilisez un token JWT Bearer dans l'en-tête `Authorization: Bearer <token>`

    ## 📋 Sections de l'API
    - **Public** : Routes accessibles sans authentification
    - **Client** : Routes pour les clients authentifiés
    - **Admin** : Routes pour les administrateurs
    - **Finance** : Routes pour l'équipe finance

  contact:
    name: Support Laine Déco
    email: support@laine-deco.cm
    url: https://laine-deco.cm
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Serveur local
  - url: https://api.laine-deco.cm
    description: Serveur de production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Entrez le token JWT
  schemas:
...
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        productCount:
          type: integer
...
security:
  - bearerAuth: []
paths:
...
  /api/public/promotions/category/{categoryId}:
    get:
      tags:
        - Public - Promotions
      summary: Promotions par catégorie
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Promotions de la catégorie
...
  /api/public/order-tracking/by-number/{orderNumber}:
    get:
      tags:
        - Public - Order Tracking
      summary: Suivi commande par numéro
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Suivi de la commande
  /api/public/order-tracking/by-id/{orderId}:
    get:
      tags:
        - Public - Order Tracking
      summary: Suivi commande par ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Suivi de la commande

  # ========== PUBLIC NEWSLETTER ROUTES ==========
...
  /api/public/maintenance-guides/{guideId}:
    get:
      tags:
        - Public - Maintenance Guides
      summary: Détails d'un guide d'entretien
      parameters:
        - name: guideId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Détails du guide
  /api/public/maintenance-guides/category/{categoryId}:
    get:
      tags:
        - Public - Maintenance Guides
      summary: Guides par catégorie
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Guides de la catégorie
  /api/public/maintenance-guides/brand/{brand}:
    get:
      tags:
        - Public - Maintenance Guides
      summary: Guides par marque
      parameters:
        - name: brand
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Guides de la marque
  /api/public/maintenance-guides/product/{productId}:
    get:
      tags:
        - Public - Maintenance Guides
      summary: Guides pour un produit
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Guides du produit
  /api/public/maintenance-guides/search:
    get:
      tags:
        - Public - Maintenance Guides
      summary: Rechercher des guides
      parameters:
        - name: search
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Résultats de recherche

  # ========== PUBLIC LEGAL PAGES ROUTES ==========
...
  /api/admin/categories:
    post:
      tags:
        - Admin - Categories
      summary: Créer une catégorie
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryResponse'
      responses:
        '201':
          description: Catégorie créée

    get:
      tags:
        - Admin - Categories
      summary: Lister les catégories
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Catégories
  /api/admin/categories/{id}:
    get:
      tags:
        - Admin - Categories
      summary: Détails catégorie
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Catégorie

    put:
      tags:
        - Admin - Categories
      summary: Mettre à jour catégorie
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryResponse'
      responses:
        '200':
          description: Catégorie mise à jour

    delete:
      tags:
        - Admin - Categories
      summary: Supprimer catégorie
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Catégorie supprimée

  # ========== ADMIN PROMOTIONS ROUTES ==========
...
  /api/admin/audit-logs/user/{userId}:
    get:
      tags:
        - Admin - Audit Logs
      summary: Logs d'audit par utilisateur
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Logs
  /api/admin/audit-logs/action/{action}:
    get:
      tags:
        - Admin - Audit Logs
      summary: Logs d'audit par action
      security:
        - bearerAuth: []
      parameters:
        - name: action
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Logs
  /api/admin/audit-logs/search:
    get:
      tags:
        - Admin - Audit Logs
      summary: Rechercher les logs d'audit
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: entityType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: startTime
          in: query
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Résultats

  # ========== FINANCE ROUTES ==========
...